Don't be smart in regex_helper, copy over core classes from template.base, rename importer so it's easier to follow
